requirements.txt
Django==1.11.17
pytz==2018.7
psycopg2-binary==2.7.4
django-filter==1.1
django-rest-swagger==2.2.0
djangorestframework==3.8.2
